validaciones

+++ playlist/src/main/java/ar/edu/unnoba/pdyc2021/mymusic/service/PlaylistServiceImp.java
@@ -3,6 +3,7 @@ package ar.edu.unnoba.pdyc2021.mymusic.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import ar.edu.unnoba.pdyc2021.mymusic.model.Song;
 import ar.edu.unnoba.pdyc2021.mymusic.model.User;
 import ar.edu.unnoba.pdyc2021.mymusic.repository.PlaylistRepository;
 import ar.edu.unnoba.pdyc2021.mymusic.repository.Playlists_SongsRepository;
+import ar.edu.unnoba.pdyc2021.mymusic.repository.SongRepository;
 import ar.edu.unnoba.pdyc2021.mymusic.repository.UserRepository;
 
 @Service
@@ -25,6 +27,8 @@ public class PlaylistServiceImp implements PlaylistService {
 	
 	@Autowired
     private Playlists_SongsRepository playlists_songsRepository;
+
+	private SongRepository songRepository;
 
     @Override
     public List<Playlist> getPlaylists() {
@@ -45,7 +49,7 @@ public class PlaylistServiceImp implements PlaylistService {
        }else{
     	   Playlist p = playlistRepository.findById(id).get();
        	 if(p.getOwner().equals(user)){
-                userRepository.deleteById(id); 
+                playlistRepository.deleteById(id); 
                 }  	
          else {
            throw new Exception("You cannot delete another playlist.");
@@ -62,19 +66,39 @@ public class PlaylistServiceImp implements PlaylistService {
 	}
 
 	@Override
-	public String addSongOnPlaylist(Playlist p, Song s) {
-		Playlists_Songs ps = new Playlists_Songs();
-		ps.setPlaylist(p);
-		ps.setSong(s);
-		playlists_songsRepository.save(ps);
-		return "";
+	public void addSongOnPlaylist(Playlist p, Song s,String loggedEmail) throws Exception{
+		User user = userRepository.findByEmail(loggedEmail);
+		if(p.getOwner().equals(user)){
+			Playlists_Songs ps = new Playlists_Songs();
+    		ps.setPlaylist(p);
+    		ps.setSong(s);
+    		if (playlists_songsRepository.findPlaylists_Songs(p, s)==null) {
+    		playlists_songsRepository.save(ps);
+    		}else {
+    			throw new Exception("The music already exists.");
+    		}
+    		}
+		else {
+			throw new Exception("Failed to add song. You are not the owner of the playlist ");
+			}
+		}
+	
 		
-	}
 
 	@Override
-	public void deleteSongOnPlaylist(Playlist p, Song s) {
-		playlists_songsRepository.delete(playlists_songsRepository.findPlaylists_Songs(p, s));
-	}
+	public void deleteSongOnPlaylist(Playlist p, Song s, String loggedEmail) throws Exception{
+		User user = userRepository.findByEmail(loggedEmail);
+		if(p.getOwner().equals(user)){
+			if(playlists_songsRepository.findPlaylists_Songs(p, s)!=null)
+			{
+			playlists_songsRepository.delete(playlists_songsRepository.findPlaylists_Songs(p, s));
+    		}else {throw new Exception("The music is not on the list.");}
+			}
+		else {
+			throw new Exception("Failed to delete song. You are not the owner of the playlist ");
+			}
+		}
+		
 
 	@Override
 	public void updatePlaylist(Long id, String name, String loggedEmail) throws Exception {
@@ -91,6 +115,11 @@ public class PlaylistServiceImp implements PlaylistService {
        }
 	}
        }
+
+	@Override
+	public boolean isExist(Long id) {
+		return playlistRepository.existsById(id);
+	}
 
 		
 	}
